[ALPS02196376] nfc: [Coverity][6735][alps-trunk-m0.tk] modification
[Detail]
NFC coverity modificaton

[Solution]
mt6605.c modification

[Feature] [Android Default] NFC for MT6605

MTK-Commit-Id: 081e16182af9a5ffcd549ff9220ac31cda94c7b9

CR-Id: ALPS02196376

+++ drivers/misc/mediatek/nfc/mt6735/mt6605.c
@@ -126,9 +126,9 @@ struct pinctrl_state *st_irq_init = NULL;
 
 /* For DMA */
 static char *I2CDMAWriteBuf;	/*= NULL;*//* unnecessary initialise */
-static unsigned int I2CDMAWriteBuf_pa;	/* = NULL; */
+static unsigned long long I2CDMAWriteBuf_pa;	/* = NULL; */
 static char *I2CDMAReadBuf;	/*= NULL;*//* unnecessary initialise */
-static unsigned int I2CDMAReadBuf_pa;	/* = NULL; */
+static unsigned long long I2CDMAReadBuf_pa;	/* = NULL; */
 
 static int fgNfcChip;		/*= 0;*//* unnecessary initialise */
 int forceExitBlockingRead = 0;
